print lists

+++ 0x12-singly_linked_lists/0-print_list.c
@@ -21,7 +21,7 @@ size_t print_list(const list_t *h)
 		}
 		else
 		{
-			printf("[%d] %s\n", h->len, h->str);
+			printf("[%lu] %s\n", (unsigned long)h->len, h->str);
 		}
 		count++;
 		h = h->next;
